+++ lol/Crypto/Lol/Cyclotomic/UCyc.hs
@@ -5,7 +5,6 @@
 {-# LANGUAGE GADTs                 #-}
 {-# LANGUAGE InstanceSigs          #-}
 {-# LANGUAGE MultiParamTypeClasses #-}
-{-# LANGUAGE NoImplicitPrelude     #-}
 {-# LANGUAGE PolyKinds             #-}
 {-# LANGUAGE RankNTypes            #-}
 {-# LANGUAGE RebindableSyntax      #-}
@@ -49,8 +48,9 @@ module Crypto.Lol.Cyclotomic.UCyc
 , coeffsPow, coeffsDec, powBasis, crtSet
 ) where
 
-import Crypto.Lol.Cyclotomic.Tensor hiding (embedCRT, embedDec, embedPow,
-                                     scalarCRT, scalarPow, twaceCRT, divGPow, divGDec)
+import Crypto.Lol.Cyclotomic.Tensor hiding (divGDec, divGPow, embedCRT,
+                                     embedDec, embedPow, scalarCRT,
+                                     scalarPow, twaceCRT)
 
 import           Crypto.Lol.CRTrans
 import           Crypto.Lol.Cyclotomic.CRTSentinel
@@ -67,7 +67,7 @@ import qualified Algebra.ZeroTestable as ZeroTestable (C)
 import Control.Applicative    as A
 import Control.Arrow
 import Control.DeepSeq
-import Control.Monad.Identity
+import Control.Monad.Identity (Identity(..))
 import Control.Monad.Random
 import Data.Foldable          as F
 import Data.Maybe
@@ -363,8 +363,9 @@ mulG (Dec v) = Dec $ mulGDec v
 mulG (CRTC s v) = CRTC s $ mulGCRTCS s v
 mulG (CRTE s v) = CRTE s $ runIdentity mulGCRT v
 
--- Note: We do not allow divGCRTE because division by g in K (might be)
--- meaningless when converted back to R.
+-- Note: We do not implement divGCRTE because we can't tell whether
+-- the element is actually divisible by g when using the CRT extension
+-- basis.
 
 -- | Divide by the special element \(g_m\).
 -- WARNING: this implementation is not a constant-time algorithm, so
@@ -375,15 +376,17 @@ divGPow :: (Fact m, UCRTElt t r, ZeroTestable r, IntegralDomain r)
 {-# INLINABLE divGPow #-}
 divGPow (Pow v) = Pow <$> T.divGPow v
 
+-- | Similar to 'divGPow'.
 divGDec :: (Fact m, UCRTElt t r, ZeroTestable r, IntegralDomain r)
         => UCyc t m D r -> Maybe (UCyc t m D r)
 {-# INLINABLE divGDec #-}
 divGDec (Dec v) = Dec <$> T.divGDec v
 
+-- | Similar to 'divGPow'.
 divGCRTC :: (Fact m, UCRTElt t r)
-        => UCyc t m C r -> Maybe (UCyc t m C r)
-{-# INLINABLE divGCRTC #-}
-divGCRTC (CRTC s v) = Just $ CRTC s $ divGCRTCS s v
+        => UCyc t m C r -> UCyc t m C r
+{-# INLINE divGCRTC #-}
+divGCRTC (CRTC s v) = CRTC s $ divGCRTCS s v
 
 -- | Yield the scaled squared norm of \(g_m \cdot e\) under
 -- the canonical embedding, namely,
@@ -547,6 +550,7 @@ crtSet =
 
 
 --------- Conversion methods ------------------
+
 -- | Convert to powerful-basis representation.
 toPow :: (Fact m, UCRTElt t r) => UCyc t m rep r -> UCyc t m P r
 {-# INLINABLE toPow #-}